Fixed bad native recording

+++ tests/test_font.cpp
@@ -51,5 +51,5 @@ BOOST_AUTO_TEST_CASE(TestCreateFont)
     BOOST_CHECK_EQUAL(10, f.size);
     BOOST_CHECK(!f.font_items[31]);
     BOOST_CHECK(f.font_items[32]);
-    BOOST_CHECK((uint32_t)f.font_items[0x4dff]);
+    BOOST_CHECK((uint64_t)f.font_items[0x4dff]);
 }
